move `uisrv` to root directory
