+++ uisrv/auth_test.go
