+++ uisrv/generic.go
