+++ uisrv/setting_test.go
